+++ Source/EventFlow.Sql/ReadModels/ReadModelSqlGenerator.cs
@@ -38,6 +38,7 @@ namespace EventFlow.Sql.ReadModels
         private static readonly ConcurrentDictionary<Type, string> TableNames = new ConcurrentDictionary<Type, string>();
         private static readonly ConcurrentDictionary<Type, IReadOnlyCollection<PropertyInfo>> PropertyInfos = new ConcurrentDictionary<Type, IReadOnlyCollection<PropertyInfo>>();
         private static readonly ConcurrentDictionary<Type, string> IdentityColumns = new ConcurrentDictionary<Type, string>();
+        private static readonly ConcurrentDictionary<Type, string> VersionColumns = new ConcurrentDictionary<Type, string>();
         private readonly Dictionary<Type, string> _insertSqls = new Dictionary<Type, string>();
         private readonly Dictionary<Type, string> _purgeSqls = new Dictionary<Type, string>();
         private readonly Dictionary<Type, string> _deleteSqls = new Dictionary<Type, string>();
@@ -102,18 +103,24 @@ namespace EventFlow.Sql.ReadModels
             where TReadModel : IReadModel
         {
             var readModelType = typeof(TReadModel);
-            string sql;
-            if (_updateSqls.TryGetValue(readModelType, out sql))
+            if (_updateSqls.TryGetValue(readModelType, out var sql))
             {
                 return sql;
             }
 
             var identityColumn = GetIdentityColumn<TReadModel>();
+            var versionColumn = GetVersionColumn<TReadModel>();
+            var versionCheck = string.IsNullOrEmpty(versionColumn)
+                ? string.Empty
+                : $"AND {versionColumn} = @_PREVIOUS_VERSION";
+
             sql = string.Format(
-                "UPDATE {0} SET {1} WHERE {2} = @{2}",
+                "UPDATE {0} SET {1} WHERE {2} = @{2} {3}",
                 GetTableName<TReadModel>(),
                 string.Join(", ", GetUpdateColumns<TReadModel>().Select(c => string.Format("{0} = @{0}", c))),
-                identityColumn);
+                identityColumn,
+                versionCheck);
+
             _updateSqls[readModelType] = sql;
 
             return sql;
@@ -155,9 +162,12 @@ namespace EventFlow.Sql.ReadModels
                 t =>
                 {
                     var tableAttribute = t.GetTypeInfo().GetCustomAttribute<TableAttribute>(false);
-                    return tableAttribute != null
-                        ? $"[{tableAttribute.Name}]"
-                        : $"[ReadModel-{t.Name.Replace("ReadModel", string.Empty)}]";
+                    var table = string.IsNullOrEmpty(tableAttribute?.Name)
+                        ? $"ReadModel-{t.Name.Replace("ReadModel", string.Empty)}"
+                        : tableAttribute.Name;
+                    return string.IsNullOrEmpty(tableAttribute?.Schema)
+                        ? $"[{table}]"
+                        : $"[{tableAttribute?.Schema}].[{table}]";
                 });
         }
 
@@ -169,6 +179,23 @@ namespace EventFlow.Sql.ReadModels
                 {
                     var propertyInfo = GetPropertyInfos(t).SingleOrDefault(pi => pi.GetCustomAttributes().Any(a => a is SqlReadModelIdentityColumnAttribute));
                     return propertyInfo?.Name ?? "AggregateId";
+                });
+        }
+
+        private string GetVersionColumn<TReadModel>()
+        {
+            return VersionColumns.GetOrAdd(
+                typeof(TReadModel),
+                t =>
+                {
+                    var propertyInfo = GetPropertyInfos(t).SingleOrDefault(pi => pi.GetCustomAttributes().Any(a => a is SqlReadModelVersionColumnAttribute));
+                    if (propertyInfo != null)
+                    {
+                        return propertyInfo.Name;
+                    }
+                    return GetPropertyInfos(t).Any(n => n.Name == "LastAggregateSequenceNumber")
+                        ? "LastAggregateSequenceNumber"
+                        : string.Empty;
                 });
         }